// samples/Dressca/dressca-frontend/consumer/src/generated/api-client/models/get-display-item-categories-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca Consumer Web API
 * Dressca Consumer の Web API 仕様
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 陳列品カテゴリのレスポンスデータを表します。
 */
export interface GetDisplayItemCategoriesResponse {
    /**
     * 陳列品カテゴリ Id を取得または設定します。
     */
    'id': string;
    /**
     * カテゴリ名を取得または設定します。
     */
    'name': string;
}