// samples/Dressca/dressca-frontend/consumer/src/generated/api-client/models/get-display-item-brands-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca Consumer Web API
 * Dressca Consumer の Web API 仕様
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 陳列品ブランドのレスポンスデータを表します。 陳列品の製造元や企画元に基づいて定義されるブランドを表現します。
 */
export interface GetDisplayItemBrandsResponse {
    /**
     * 陳列品ブランド Id を取得または設定します。
     */
    'id': string;
    /**
     * ブランド名を取得または設定します。
     */
    'name': string;
}