// samples/Dressca/dressca-frontend/consumer/src/generated/api-client/models/paged-list-of-get-display-item-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca Consumer Web API
 * Dressca Consumer の Web API 仕様
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { GetDisplayItemResponse } from './get-display-item-response';

/**
 * ページネーションした結果のリストを管理します。
 */
export interface PagedListOfGetDisplayItemResponse {
    /**
     * 現在のページ番号を取得します。 最初のページは 1 です。
     */
    'page'?: number;
    /**
     * ページの総数を取得します。
     */
    'totalPages'?: number;
    /**
     * 1 ページあたりの件数を取得します。
     */
    'pageSize'?: number;
    /**
     * アイテムの総数を取得します。
     */
    'totalCount'?: number;
    /**
     * 前のページが存在するかどうか示す値を取得します。
     */
    'hasPrevious'?: boolean;
    /**
     * 次のページが存在するかどうか示す値を取得します。
     */
    'hasNext'?: boolean;
    /**
     * 検索結果のリストを取得します。
     */
    'items'?: Array<GetDisplayItemResponse>;
}